backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
python-multipart==0.0.6
starlette==0.27.0
jinja2==3.1.2
websockets==11.0.3
sqlalchemy==2.0.23
sse-starlette==1.6.5
bcrypt==4.1.2 